cttseraser: use statx() instead of stat()
stat() fails on files with a date after 2038 on 32 bit systems
unless _TIME_BITS=64 is defined when building, however that option
appears to be incompatible with FUSE, as it results in all file
sizes being 16384, likely due to incompatible struct stat layout.
Use statx() instead, then copy the fields from the struct statx to
the struct stat, clamping timestamps to 2038 if necessary.